+++ ios/WonderousWidget/Colors.swift
@@ -1,0 +1,9 @@
+import Foundation
+import SwiftUI
+
+struct Colors {
+    static let accentColor:Color = Color(red: 0.89, green: 0.58, blue: 0.36)
+    static let offWhiteColor:Color = Color(red: 0.97, green: 0.92, blue: 0.9)
+    static let mediumGrey:Color = Color(red: 0.62, green: 0.6, blue: 0.58)
+    static let darkGrey:Color = Color(red: 0.32, green: 0.31, blue: 0.3);
+}
