+++ ios/WonderousWidget/WonderousWidget.swift
@@ -1,22 +1,10 @@
-//
-//  CounterWidget.swift
-//  CounterWidget
-//
-//  Created by Shawn on 2023-09-11.
-//
-
 import WidgetKit
 import SwiftUI
 import Intents
 
-var netImgData: Data? = nil
-
-
 /// Entry, is passed into the view and defines the data it needs
 struct WonderousEntry : TimelineEntry {
     let date: Date
-    //let displaySize: CGSize
-    //let imageData: Data?
     let discoveredCount:Int;
     var title:String = "";
     var subTitle:String = "";
@@ -24,8 +12,7 @@ struct WonderousEntry : TimelineEntry {
 
 }
 
-
-// Widget, defines the display name and description, and also wraps the View
+// Widget, defines the display name and description and also declared the main View
 struct WonderousWidget: Widget {
     let kind: String = "WonderousWidget"
     
@@ -69,14 +56,6 @@ struct Provider: TimelineProvider {
     
     // Provide an array of entries for the current time and, optionally, any future times
     func getTimeline(in context: Context, completion: @escaping (Timeline<WonderousEntry>) -> ()) {
-        // Load a remote image so it can be shown later
-//        let userDefaults = UserDefaults(suiteName: "group.com.gskinner.flutter.wonders.widget")
-//        let url = userDefaults?.string(forKey: "lastDiscoveredImageUrl");
-//        if(url != nil){
-//            netImgData = try? Data(contentsOf: URL(string: url!)!)
-//        } else {
-//            netImgData = nil;
-//        }
         getSnapshot(in: context) { (entry) in
             let timeline = Timeline(entries: [entry], policy: .atEnd)
             completion(timeline)
